backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
supabase-py==2.3.5
pdfkit==1.0.0
python-docx==1.1.0